requirements/all.txt
Django==1.7
argparse==1.2.1
django-filter==0.8
djangorestframework==2.4.3
ipdb==0.8
ipython==2.2.0
konch==0.4.2
psycopg2==2.5.4
six==1.8.0
wsgiref==0.1.2
